have back end connected and can make db request
