+++ index.html
@@ -1,0 +1,178 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons"
+      rel="stylesheet">
+    <link rel="stylesheet" href="/stylesheets/style.css">
+</head>
+
+  <body>
+    <header>
+        <h1 class="appTitle">Weather App</h1>
+    </header>
+
+    <div class="formWrapper">
+        <form method="POST" action="/searchcity" id="searchCityForm">
+
+            <div class="formInputWrapper">
+                <button id="currLocationBtn"><span class="material-icons">my_location</span></button>
+                <input type="text" name="searchCityTextInput" id="searchCityTextInput" placeholder="Search City" autocomplete="off" required>
+                <input type="text" name="searchCountryTextInput" id="searchCountryTextInput" placeholder="US" maxlength="2" size="1" value="US" autocomplete="off" required>
+            </div>
+
+            <div class="searchResultsWrapperMain">
+                <div class="searchResultsWrapper">
+                    <ul id=citySearchResults></ul>
+                    <ul id="countrySearchResults"></ul>
+                </div>
+            </div>
+
+        </form>
+    </div>
+
+    <!-- if API call error, or any errors happened-->
+    <p id="errorMessage" style="text-align: center;"></p>
+
+    <!-- div will display all weather information for selected city-->
+    <div class="dataWrapperMain">
+        <div id="dataWrapper">
+            <p id="cityName"></p>
+            <img id="weatherImg">
+            <p id=currentTemp></p>
+            <button id="tempUnitBtn">F/C</button>
+        </div>
+    </div>
+
+    <footer>
+        <div class="footerWrapper">
+            <div class="footerBottom" style="text-align: center;">Weather App</div>
+        </div>
+    </footer>
+
+  </body>
+</html>
+
+<!---------------------------- Start Javascript ------------------------------>
+
+<!-- code block handles city and country search -->
+<script type="module">
+    import {queryCity, queryCountry} from "./jsmodules/querylocations.js";
+    
+    let searchCityTextInput = document.getElementById('searchCityTextInput');
+    let searchList = document.getElementById('citySearchResults');
+    let searchCountryTextInput = document.getElementById('searchCountryTextInput');
+    let countryList = document.getElementById('countrySearchResults');
+    let reAlpha = /^[a-z]+$/i;
+    
+    // Search city text input listener event
+    searchCityTextInput.addEventListener('input', (event) => {
+        // only letters a-z will trigger database query
+        if(reAlpha.test(event.data)) {
+            queryCity(searchCityTextInput, searchCountryTextInput, searchList);
+        }
+        return;
+    });
+    // Search country text input
+    searchCountryTextInput.addEventListener('input', (event) => {
+        
+        // remove search list for cities
+        clearSearchResults(searchList)
+        // only letters a-z will trigger database query
+        if(reAlpha.test(event.data)) {
+            queryCountry(searchCountryTextInput, searchList);
+        };
+    });
+</script>
+
+
+<!-- Function removes search result list from the DOM -->
+<script>
+    let clearSearchResults = (list) => {
+        // check to see if DOM list contains anything, if so, remove them
+        if(list.childElementCount > 0) {
+            let first = list.firstElementChild;
+            while(first) {
+                first.remove();
+                first = list.firstElementChild;
+            };
+        };
+    };
+</script>
+
+<!--------------- API call and display weather results ---------------->
+<script>
+    // variables for view data
+    let errorMessage = document.getElementById('errorMessage');
+    let cityName = document.getElementById('cityName');
+    let currentTemp = document.getElementById('currentTemp');
+    let dataWrapper = document.getElementById('dataWrapper');
+    let weatherImg = document.getElementById('weatherImg');
+    let fahrenheit;
+</script>
+
+
+<!-- Function converts temp between fahrenheit/cel with button -->
+<script>
+    let convertTemp = (kelvin) => {
+    
+    switch(fahrenheit) {
+        // for celcius
+        case true: 
+            fahrenheit = false;
+            currentTemp.textContent = Math.round(kelvin - 273.15) + " C";
+            return;
+        // for fahrenheit
+        case false:
+            fahrenheit = true;
+            currentTemp.textContent = Math.round((kelvin - 273.15) * 1.8 + 32) + " F";
+            return; 
+    };
+};
+</script>
+
+
+<!-- Function displays city weather to the DOM -->
+<script>
+    //TODO: display more data. e.g. high and low temp, wind, sunrise/sunset times
+    let displayWeatherData = (weatherData) => {
+    console.log(weatherData);
+    // show the data div
+    dataWrapper.style.display = "flex";
+    cityName.textContent = weatherData.name;
+    weatherImg.src = `http://openweathermap.org/img/wn/${weatherData.weather[0].icon}@2x.png`;
+    // default display temp in fahrenheit
+    fahrenheit = false;
+    convertTemp(weatherData.main.temp);
+    
+    // toggle temp unit
+    document.getElementById('tempUnitBtn').onclick = () => {
+        convertTemp(weatherData.main.temp);      
+    };
+};
+</script>
+
+<!-- function makes api call to get weather data -->
+<script>
+    async function fetchApi(e) {
+        // get input from the form
+        let cityId = e.target.dataset.cityid;
+        // api call
+        let response = await fetch(`https://api.openweathermap.org/data/2.5/weather?id=${cityId}&appid=65388c50a787be295df1ae5b1f2c37ea`, {mode: 'cors'});
+        let weatherData = await response.json();
+        // handle errors
+        if(!response.ok) {
+            errorMessage.textContent = data.message;
+            dataWrapper.style.display = "";
+            searchCityTextInput.value = "";
+            // TODO send error message to DOM
+            return;
+        };
+        
+        searchCityTextInput.value = ""
+        return(displayWeatherData(weatherData))
+    };
+</script>
