+++ weather-app-backend/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "better-sqlite3": "^7.5.0",
     "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
     "debug": "^4.3.3",
     "express": "^4.17.2",
     "http-errors": "^2.0.0",
