+++ weather-app-backend/routes/index.js
@@ -1,6 +1,18 @@
 var express = require('express');
 var router = express.Router();
+var app = express();
 const index = require('../controllers/index')
+
+
+// Allow requests from vue front end
+const allowRequest = app.use(function(req, res, next) {
+    
+    res.header("Access-Control-Allow-Origin", "*");
+    // res.header("Access-Control-Allow-Methods: GET, POST, OPTIONS, PUT, DELETE");
+    // res.header("Access-Control-Allow-Headers: Authorization");
+    // res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    next();
+});
 
 router.get('/', index.index);
 
@@ -8,5 +20,6 @@ router.get('/', index.index);
 router.post('/searchcity', index.searchCity);
 
 // POST route for country search
-router.post('/searchcountry', index.searchCountry);
+router.post('/searchcountry', allowRequest, index.searchCountry);
+
 module.exports = router;
