+++ weather-app/src/views/Home.vue
@@ -8,6 +8,7 @@
 
 <script>
 // @ is an alias to /src
+import axios from 'axios'
 import Header from '@/components/Header.vue'
 import Searchbar from '@/components/Searchbar.vue'
 import Footer from '@/components/Footer.vue'
@@ -18,6 +19,20 @@ export default {
     Header,
     Searchbar,
     Footer
+  },
+  data() {
+      
+      return {
+          info: null
+      }
   }
+//   mounted() {
+
+//       axios.get('http://127.0.0.1:3000/searchcountry')
+//         .then((response) => {
+//             this.info = response;
+//         }) 
+//   }
+  
 }
 </script>
